Added Player class, noted desired usage
Added some clarification notes in player.py for how the object should
ideally interact with the game board.

Noting that the game board should be a subclass of HexGrid, with the sie
and colors already specified.

Experimented with the timer fucntion, results yielded by pressing
key.SPACE to start the timer are seemingly erratic time keeping. Will
need higher precision in that area.